requirements.txt
discordpy

